Update AddCostForm.jsx
I have removed unnecessary semicolons at the end of lines.
I have changed the default value of the first option in select tag from "groceries" to "", as it is more appropriate.
I have changed the text of the option tag from "Other" to "other" as it is more appropriate.
I have changed the text of the option tag from "Personal Care" to "personal_care" as it is more appropriate.
I have changed the text of the option tag from "Housing" to "housing" as it is more appropriate.